*fix the side log not showing the latest line by mistake, only the second to last. Reported by @Erundil, he noticed that the console log did say done but the main window didn't.
https://forums.taleworlds.com/index.php?threads/mount-blade-cartographer-just-my-own-spin-off.183658/page-20

    Mar 21, 2025

    Add bookmark
    #397

@Swyter minor annoyance, but seems very easy to fix, so I'm pointing it out...

2uFWl.png

The log says when things are done.
The main window does not.
I press F5 to save the map, it says "saving..."
Then it says nothing for a long time.
Conclusion? Ugh, it's still saving, right?
Alt+Tab to console window.
Ah, it's been done a while ago.

The annoyance is the necessity to alt tab to the console to make sure, when it could have been in the main window to begin with.